api/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pillow==10.4.0
opencv-python-headless==4.10.0.84
numpy==2.0.2
